+++ src/Credfeto.Dispatcher.GitHub/Configuration/GitHubFilterOptions.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Diagnostics;
-using Credfeto.Dispatcher.GitHub.DataTypes;
 
 namespace Credfeto.Dispatcher.GitHub.Configuration;
 
@@ -25,5 +24,5 @@ public sealed class GitHubFilterOptions
 
     public string MentionedUser { get; set; } = string.Empty;
 
-    public IReadOnlyList<BotPrRule> BotPrRules { get; set; } = [];
+    public GitHubPullRequestFilterOptions PullRequests { get; set; } = new();
 }